requirements.txt
dify_plugin==0.5.0b15
markdown-it-py>=3.0.0
pydantic>=2.0.0
PyYAML>=6.0.0
pytest>=8.0.0
hypothesis>=6.0.0

# Core markdown processing dependencies
mistune>=3.0.0
markdown>=3.4.0
mdformat>=0.7.0
markdown2>=2.4.0
dataclasses-json>=0.6.0

# Development dependencies (optional, for code quality)
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
